requirements_old.txt
# certifi                2023.6.8
# charset-normalizer     3.3.2
# idna                   3.4
mysql-connector        2.2.9
mysql-connector-python 8.2.0
numpy                  1.26.2
pandas                 2.1.3
# pip                    22.0.2
protobuf               4.21.12
# pynytimes              0.10.0
python-dateutil        2.8.2
pytz                   2023.3.post1
requests               2.31.0
setuptools             59.6.0
six                    1.16.0
tzdata                 2023.3
urllib3                2.1.0
